Permission checks in functions

+++ dev/ServiceLayer/TransportServices/SiteService.java
@@ -18,15 +18,13 @@ public class SiteService {
 
 
     public String addShippingArea(long loggedID, int areaNum, String areaName){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "ADD_SHIPPING_AREA")){
+            return "You are not authorized to make this action !";
+        }
         if (!this.employeeIntegrationService.isActive(loggedID)){ return "You are not an active employee, you can't add Shipping Areas"; }
         if (!this.employeeIntegrationService.hasRole(loggedID, "Admin") && !this.employeeIntegrationService.hasRole(loggedID, "Transport manager")){
             return "You don't have the required role to add Shipping Areas, you can only add Shipping Areas if you are an Admin or a Transport Manager";
         }
-        //TODO:  We need to add a Permission checking function to the EmployeeIntegrationService.
-        //TODO:  We need to add a Permission checking function to the EmployeeIntegrationService.
-        //TODO:  We need to add a Permission checking function to the EmployeeIntegrationService.
-        //TODO:  We need to add a Permission checking function to the EmployeeIntegrationService.
-        //TODO:  We need to add a Permission checking function to the EmployeeIntegrationService.
         try {
             if(areaName.isEmpty() || areaName.isBlank()){
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -42,6 +40,9 @@ public class SiteService {
     }
 
     public String deleteShippingArea(long loggedID, int areaNum){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "DELETE_SHIPPING_AREA")){
+            return "You are not authorized to make this action !";
+        }
         try {
             sf.deleteShippingArea(areaNum);
         } catch (AttributeNotFoundException e){
@@ -58,6 +59,9 @@ public class SiteService {
 
 
     public String setShippingAreaNum(long loggedID, int OldareaNum, int NewAreaNum){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SHIPPING_AREA")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if(OldareaNum == NewAreaNum){
                 return "The Edition process Finished because you set the same Area Number value as the value that is already there";
@@ -76,6 +80,9 @@ public class SiteService {
 
 
     public String setShippingAreaName(long loggedID, int areaNum, String NewareaName){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SHIPPING_AREA")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if(NewareaName.isEmpty() || NewareaName.isBlank()){
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -106,6 +113,9 @@ public class SiteService {
 
 
     public String addSite(long loggedID, int areaNum, String address, String cont_name, long Cont_Num){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "ADD_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (address.isEmpty() || cont_name.isEmpty() || address.isBlank() || cont_name.isBlank() || Cont_Num == 0) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -123,6 +133,9 @@ public class SiteService {
     }
 
     public String deleteSite(long loggedID, int areaNum, String address){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "DELETE_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (address.isEmpty() || address.isBlank()) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -144,6 +157,9 @@ public class SiteService {
 
 
     public String setSiteAddress(long loggedID, int areaNum, String Oldaddress, String NewAddress){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (NewAddress.isEmpty() || Oldaddress.isEmpty() || NewAddress.isBlank() || Oldaddress.isBlank()) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -163,6 +179,9 @@ public class SiteService {
     }
 
     public String setSiteAreaNum(long loggedID, int OldareaNum, int NewAreaNum, String address){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (address.isEmpty() || address.isBlank()) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -182,6 +201,9 @@ public class SiteService {
     }
 
     public String setSiteContName(long loggedID, int areaNum, String address, String contName){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (address.isEmpty() || contName.isEmpty() || address.isBlank() || contName.isBlank()) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -197,6 +219,9 @@ public class SiteService {
     }
 
     public String setSiteContNum(long loggedID, int areaNum, String address, long contNum){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "EDIT_SITE")){
+            return "You are not authorized to make this action !";
+        }
         try {
             if (address.isEmpty() || address.isBlank()) {
                 return "It seems some values you've entered are Empty or Blank, Please Insert proper values";
@@ -226,6 +251,9 @@ public class SiteService {
 
 
     public String showAllSites(long loggedID){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "SHOW_SITES")){
+            return "You are not authorized to make this action !";
+        }
         String res = "";
         try {
             res = sf.showAllSites();
@@ -236,6 +264,9 @@ public class SiteService {
     }
 
     public String showAllShippingAreas(long loggedID){
+        if (!this.employeeIntegrationService.isEmployeeAuthorised(loggedID, "SHOW_SHIPPING_AREAS")){
+            return "You are not authorized to make this action !";
+        }
         String res = "";
         try {
             res = sf.showAllShippingAreas();
